Use AdamW with lr_ratio setting instead of AdamWDL
