+++ model_zoo/ernie-doc/run_mrc.py
@@ -24,6 +24,7 @@ import time
 import numpy as np
 import paddle
 import paddle.nn as nn
+from paddle.optimizer import AdamW
 from paddle.io import DataLoader
 from paddlenlp.transformers import ErnieDocModel
 from paddlenlp.transformers import ErnieDocForQuestionAnswering
@@ -31,7 +32,7 @@ from paddlenlp.transformers import ErnieDocTokenizer
 from paddlenlp.transformers import LinearDecayWithWarmup
 from paddlenlp.utils.log import logger
 from paddlenlp.datasets import load_dataset
-from paddlenlp.ops.optimizer import AdamWDL
+from paddlenlp.ops.optimizer import layerwise_lr_decay
 
 from data import MRCIterator
 from metrics import compute_qa_predictions, EM_AND_F1
@@ -253,13 +254,14 @@ def do_train(args):
     for n, p in model.named_parameters():
         name_dict[p.name] = n
 
-    optimizer = AdamWDL(learning_rate=lr_scheduler,
-                        parameters=model.parameters(),
-                        weight_decay=args.weight_decay,
-                        apply_decay_param_fun=lambda x: x in decay_params,
-                        n_layers=model_config["num_hidden_layers"],
-                        layerwise_decay=args.layerwise_decay,
-                        name_dict=name_dict)
+    simple_lr_setting = partial(layerwise_lr_decay, args.layerwise_decay,
+                                name_dict, model_config["num_hidden_layers"])
+
+    optimizer = AdamW(learning_rate=lr_scheduler,
+                      parameters=model.parameters(),
+                      weight_decay=args.weight_decay,
+                      apply_decay_param_fun=lambda x: x in decay_params,
+                      lr_ratio=simple_lr_setting)
 
     global_steps = 0
     create_memory = partial(init_memory, args.batch_size, args.memory_length,
