+++ model_zoo/ernie-doc/run_semantic_matching.py
@@ -26,6 +26,7 @@ import paddle
 import paddle.nn as nn
 from paddle.io import DataLoader
 from paddle.metric import Accuracy
+from paddle.optimizer import AdamW
 
 from paddlenlp.transformers import ErnieDocModel
 from paddlenlp.transformers import ErnieDocForSequenceClassification
@@ -33,7 +34,7 @@ from paddlenlp.transformers import ErnieDocTokenizer
 from paddlenlp.transformers import LinearDecayWithWarmup
 from paddlenlp.utils.log import logger
 from paddlenlp.datasets import load_dataset
-from paddlenlp.ops.optimizer import AdamWDL
+from paddlenlp.ops.optimizer import layerwise_lr_decay
 
 from data import SemanticMatchingIterator
 from model import ErnieDocForTextMatching
@@ -238,13 +239,14 @@ def do_train(args):
     for n, p in model.named_parameters():
         name_dict[p.name] = n
 
-    optimizer = AdamWDL(learning_rate=lr_scheduler,
-                        parameters=model.parameters(),
-                        weight_decay=args.weight_decay,
-                        apply_decay_param_fun=lambda x: x in decay_params,
-                        n_layers=model_config["num_hidden_layers"],
-                        layerwise_decay=args.layerwise_decay,
-                        name_dict=name_dict)
+    simple_lr_setting = partial(layerwise_lr_decay, args.layerwise_decay,
+                                name_dict, model_config["num_hidden_layers"])
+
+    optimizer = AdamW(learning_rate=lr_scheduler,
+                      parameters=model.parameters(),
+                      weight_decay=args.weight_decay,
+                      apply_decay_param_fun=lambda x: x in decay_params,
+                      lr_ratio=simple_lr_setting)
 
     criterion = paddle.nn.loss.CrossEntropyLoss()
     metric = paddle.metric.Accuracy()
