+++ paddlenlp/ops/optimizer/__init__.py
@@ -13,7 +13,30 @@
 # limitations under the License.
 import os
 
-from .adamwdl import AdamWDL
 from .ema import ExponentialMovingAverage
 
-__all__ = ['AdamWDL', 'ExponentialMovingAverage']
+__all__ = ['layerwise_lr_decay', 'ExponentialMovingAverage']
+
+
+# Layerwise decay
+def layerwise_lr_decay(decay_rate, name_dict, n_layers, param):
+    """
+    Args:
+        decay_rate (float): 
+            The layer-wise decay ratio.
+        name_dict (dict): 
+            The keys of name_dict is dynamic name of model while the value
+            of name_dict is static name.
+            Use model.named_parameters() to get name_dict.
+        n_layers (int):
+            Total number of layers in the transformer encoder.
+    """
+    ratio = 1.0
+    static_name = name_dict[param.name]
+    if "encoder.layers" in static_name:
+        idx = static_name.find("encoder.layers.")
+        layer = int(static_name[idx:].split(".")[2])
+        ratio = decay_rate**(n_layers - layer)
+    elif "embedding" in static_name:
+        ratio = decay_rate**(n_layers + 1)
+    return ratio
